+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for specifying multiple area, service, resource values for the respective flags.
 - Support for resources without quota.
 - Support for relative quota adjustment. The following operators are supported: `+=`,
   `-=`, `*=`, `/=`.